+++ GRDB/Core/Configuration.swift
@@ -43,6 +43,38 @@ public struct Configuration {
     /// Default: deferred
     public var defaultTransactionKind: Database.TransactionKind = .deferred
     
+    /// If false, it is a programmer error to leave a transaction opened at the
+    /// end of a database access block.
+    ///
+    /// For example:
+    ///
+    ///     let dbQueue = DatabaseQueue()
+    ///
+    ///     // fatal error: A transaction has been left opened at the end of a database access
+    ///     try dbQueue.inDatabase { db in
+    ///         try db.beginTransaction()
+    ///     }
+    ///
+    /// If true, one can leave opened transaction at the end of database access
+    /// blocks:
+    ///
+    ///     var config = Configuration()
+    ///     config.allowsUnsafeTransactions = true
+    ///     let dbQueue = DatabaseQueue(configuration: config)
+    ///
+    ///     try dbQueue.inDatabase { db in
+    ///         try db.beginTransaction()
+    ///     }
+    ///
+    ///     try dbQueue.inDatabase { db in
+    ///         try db.commit()
+    ///     }
+    ///
+    /// This configuration flag has no effect on DatabasePool readers: those
+    /// never allow leaving a transaction opened at the end of a read access.
+    ///
+    /// Default: false
+    public var allowsUnsafeTransactions: Bool = false
     
     // MARK: - Concurrency